Refactored hardware factories
- Default factories for Android, macOS, Windows
- Enhanced hardware configuration and schema
