+++ sources/hardware/hw_config.h
@@ -1,6 +1,6 @@
 /******************************************************************************
  *
- * Copyright (C) 2019 Markus Zehnder <business@markuszehnder.ch>
+ * Copyright (C) 2019-2020 Markus Zehnder <business@markuszehnder.ch>
  *
  * This file is part of the YIO-Remote software project.
  *
@@ -22,14 +22,19 @@
 
 #pragma once
 
+/**
+ * Common hardware configuration keys.
+ * Platform specific keys should be defined either in the specific factory or in a dedicated header file.
+ */
+
 #define HW_CFG_PATH_I2C_DEV       "i2c/device"
 #define HW_CFG_PATH_I2C_ID        "i2c/id"
 
 #define HW_CFG_PATH_GPIO_PIN      "gpio/pin"
 #define HW_CFG_PATH_INTR_GPIO_PIN "interrupt/gpio/pin"
 
-#define HW_CFG_DISPLAY            "display"
-#define HW_CFG_IOEXPANDER         "ioExpander"
+#define HW_CFG_DISPLAY_CONTROL    "display/control"
+#define HW_CFG_BTN_INTR_HANDLER   "buttonInterruptHandler"
 #define HW_CFG_BATTERY_CHARGER    "batteryCharger"
 #define HW_CFG_BATTERY_FUEL_GAUGE "batteryFuelGauge"
 #define HW_CFG_HAPTIC_MOTOR       "hapticMotor"
