+++ sources/hardware/linux/arm/apds9960light.cpp
