+++ sources/hardware/linux/arm/displaycontrol_yio.cpp
@@ -39,14 +39,14 @@
 
 static Q_LOGGING_CATEGORY(CLASS_LC, "hw.dev.display");
 
-DisplayControlYio::DisplayControlYio(int pin, QObject *parent)
+DisplayControlYio::DisplayControlYio(int backlightPin, QObject *parent)
     : DisplayControl("YIO display control", parent) {
-    Q_ASSERT(pin);
-    qCDebug(CLASS_LC()) << name() << "on GPIO" << pin;
+    Q_ASSERT(backlightPin);
+    qCDebug(CLASS_LC()) << name() << "[backlightPin=" << backlightPin << "]";
 
     // move the low level hardware handling to a separate thread
     m_thread = new QThread(this);
-    DisplayControlYioThread *dcyt = new DisplayControlYioThread(pin);
+    DisplayControlYioThread *dcyt = new DisplayControlYioThread(backlightPin);
 
     connect(this, &DisplayControlYio::enterStandby, dcyt, &DisplayControlYioThread::enterStandby);
     connect(this, &DisplayControlYio::leaveStandby, dcyt, &DisplayControlYioThread::leaveStandby);
@@ -130,8 +130,8 @@ void DisplayControlYioThread::setBrightness(int from, int to) {
         to = 100;
     }
 
-    if (from == 0 && digitalRead(m_pin) == 0) {
-        pinMode(m_pin, PWM_OUTPUT);
+    if (from == 0 && digitalRead(m_backlightPin) == 0) {
+        pinMode(m_backlightPin, PWM_OUTPUT);
         pwmSetMode(PWM_MODE_MS);
         pwmSetClock(1000);
         pwmSetRange(100);
@@ -144,18 +144,18 @@ void DisplayControlYioThread::setBrightness(int from, int to) {
     if (from >= to) {
         // dim down
         for (int i = from; i > to - 1; i--) {
-            pwmWrite(m_pin, i);
+            pwmWrite(m_backlightPin, i);
             delay(10);
             if (i == 0) {
                 delay(100);
-                pinMode(m_pin, OUTPUT);
-                digitalWrite(m_pin, 0);
+                pinMode(m_backlightPin, OUTPUT);
+                digitalWrite(m_backlightPin, 0);
             }
         }
     } else {
         // dim up
         for (int i = from; i < to + 1; i++) {
-            pwmWrite(m_pin, i);
+            pwmWrite(m_backlightPin, i);
             delay(10);
         }
     }
