+++ sources/hardware/linux/arm/drv2605.h
@@ -41,7 +41,7 @@
 #include <wiringPi.h>
 #include <wiringPiI2C.h>
 
-#include "../hapticmotor.h"
+#include "../../hapticmotor.h"
 
 #define DRV2605_ADDR 0x5A  ///< Device I2C address
