+++ sources/hardware/linux/webserver_lighttpd.h
@@ -22,37 +22,41 @@
 
 #pragma once
 
-#include <QMap>
-#include <QObject>
-
-#include "systemservice.h"
+#include "../systemservice.h"
+#include "../webserver_control.h"
 
 /**
- * @brief Linux systemd implementation of the SystemService interface.
+ * @brief A lighttpd implementation of the WebServerControl interface.
  */
-class Systemd : public SystemService {
+class WebServerLighttpd : public WebServerControl {
     Q_OBJECT
 
  public:
-    explicit Systemd(const QMap<SystemServiceName, QString> &serviceNameMap, QObject *parent = nullptr);
-
-    void setUseSudo(bool sudo);
-    bool isUseSudo();
-
-    // SystemService interface
- public:
-    Q_INVOKABLE bool startService(SystemServiceName serviceName) override;
-    Q_INVOKABLE bool stopService(SystemServiceName serviceName) override;
-    Q_INVOKABLE bool restartService(SystemServiceName serviceName) override;
-    Q_INVOKABLE bool reloadService(SystemServiceName serviceName) override;
-
-    int  systemctlTimeout() const;
-    void setSystemctlTimeout(int systemctlTimeout);
+    explicit WebServerLighttpd(SystemService *systemService, QObject *parent = nullptr);
 
  private:
-    bool launch(const QString &command);
+    SystemService *p_systemService;
 
-    bool                             m_useSudo;
-    int                              m_systemctlTimeout;
-    QMap<SystemServiceName, QString> m_serviceNameMap;
+    // WebServerControl interface
+ public:
+    Q_INVOKABLE bool startService() override;
+    Q_INVOKABLE bool stopService() override;
+    Q_INVOKABLE bool restartService() override;
+    Q_INVOKABLE bool reloadService() override;
+    Q_INVOKABLE bool startWifiSetupPortal() override;
+    Q_INVOKABLE bool startWebConfigurator() override;
+
+    QString configFile() const;
+    void    setConfigFile(const QString &configFile);
+
+    QString wifiSetupConfig() const;
+    void    setWifiSetupConfig(const QString &wifiSetupConfig);
+
+    QString webConfiguratorConfig() const;
+    void    setWebConfiguratorConfig(const QString &webConfiguratorConfig);
+
+ private:
+    QString m_configFile;
+    QString m_wifiSetupConfig;
+    QString m_webConfiguratorConfig;
 };
